+++ todo_app/data/item.py
@@ -1,9 +1,11 @@
 class Item: 
-    def __init__(self, id, title, status = 'To Do'): 
+    def __init__(self, idList, id, name, shortLink, status = 'To Do'): 
+        self.idList = idList 
         self.id = id
-        self.title = title 
+        self.name = name 
+        self.shortLink = shortLink
         self.status = status
    
     @classmethod 
     def from_trello_card(cls, card, list): 
-        return cls(card['id'], card['name'],  list['name'])
+        return cls(card['idList'], card['id'],card['name'], card['shortLink'], list['name'])
