+++ todo_app/templates/index.html
@@ -9,7 +9,7 @@
 
   <div class="row">
     <div class="col-md-12">
-      <h2>Todo Items</h2>
+      <h2>Items</h2>
       <ul class="list-group mb-2">
         <form action="/" method="GET">
         <!-- List the todo items here -->
